Initial commit.
Original mod by Pizzaoverhead.  Building upon the code he has already created.

Added x, y and z axis for position to the joystick controller
Fixed issue with the config file not loading or saving correctly.
Tweak to GUI adding buttons to set the x,y z axis
Tweak to the offset slider for better control over the positioning of the head when using the Joystick method